==> notebooks/1_learn_files.ipynb <==
from pathlib import Path
import re
import pandas as pd

val_list_file = Path('..') / 'data' / 'train' / 'validation_list.txt'
test_list_file = Path('..') / 'data' / 'train' / 'testing_list.txt'

lines = []

with open(val_list_file, 'r') as file:
    for line in file:
        lines.append(re.split(r'[/_\.]', line[:-1]))

val_df = pd.DataFrame(lines, columns=['folder', 'person', 'nohash', 'person_ind', 'format'])
val_df['person_ind'] = val_df['person_ind'].astype(int)

lines = []

with open(test_list_file, 'r') as file:
    for line in file:
        lines.append(re.split(r'[/_\.]', line[:-1]))

test_df = pd.DataFrame(lines, columns=['folder', 'person', 'nohash', 'person_ind', 'format'])
test_df['person_ind'] = test_df['person_ind'].astype(int)

val_df.shape, test_df.shape

val_df.sample(5)

val_df.nunique().sort_values()

test_df.nunique().sort_values()

from pathlib import Path

# Create a Path object for the root folder
root_folder = Path('..') / 'data' / 'train' / 'audio'

lines = []

for path in root_folder.rglob('*'):
    if '_background_noise_' not in path.parts and path.is_file():
        if len(re.split(r'[\\/_\.]', str(path.relative_to(root_folder)))) > 5:
            print(re.split(r'[\\/_\.]', str(path.relative_to(root_folder))))
        lines.append(re.split(r'[\\/_\.]', str(path.relative_to(root_folder))))

# # Use the rglob method to find all files in the folder and its subfolders
# file_paths = [str(path.relative_to(root_folder)) for path in root_folder.rglob('*') if path.is_file()]

# # Now file_paths is a list of all file paths, starting from the directory name inside 'root_folder'
# for path in file_paths:
#     print(path)
    
# lines = []

# with open(test_list_file, 'r') as file:
#     for line in file:
#         lines.append(re.split(r'[/_\.]', line[:-1]))

# print(lines)

full_df = pd.DataFrame(lines, columns=['folder', 'person', 'nohash', 'person_ind', 'format'])
full_df.shape

full_df['person_ind'] = full_df['person_ind'].astype(int)

full_df.nunique().sort_values()

train_df = pd.concat([full_df, val_df, test_df]).drop_duplicates(keep=False)
assert len(full_df) == len(train_df) + len(val_df) + len(test_df)
train_df.shape, val_df.shape, test_df.shape, full_df.shape

assert full_df['nohash'].unique() == ['nohash']
assert full_df['format'].unique() == ['wav']

train_df.drop(columns=['nohash', 'format'], inplace=True)
val_df.drop(columns=['nohash', 'format'], inplace=True)
test_df.drop(columns=['nohash', 'format'], inplace=True)
full_df.drop(columns=['nohash', 'format'], inplace=True)

full_df.info()

assert set(val_df['folder']) == set(test_df['folder'])
assert set(val_df['folder']) == set(full_df['folder'])

assert set(val_df['person']) & set(test_df['person']) == set()
assert set(val_df['person']) & set(train_df['person']) == set()
assert set(test_df['person']) & set(train_df['person']) == set()

# Full investigation

full_df.info()

full_df.groupby(by='folder').count()['person'].sort_values()

folders_aux = set(full_df.groupby(by='folder').count()['person'][full_df.groupby(by='folder').count()['person'] < 2000].index)
print(len(folders_aux), folders_aux)
folders_main = set(full_df.groupby(by='folder').count()['person'][full_df.groupby(by='folder').count()['person'] > 2000].index)
print(len(folders_main), folders_main)

full_aux_df = full_df[full_df['folder'].isin(folders_aux)]
full_main_df = full_df[full_df['folder'].isin(folders_main)]
full_df.shape, full_main_df.shape, full_aux_df.shape 

full_df['person'].nunique(), full_aux_df['person'].nunique(), full_main_df['person'].nunique()

full_aux_df.groupby(by='folder').nunique()['person']

full_main_df.groupby(by='folder').nunique()['person']

pd.DataFrame({'full': full_df['person_ind'].describe(),
              'aux': full_aux_df['person_ind'].describe(),
              'main': full_main_df['person_ind'].describe()})

train_main_df = train_df[train_df['folder'].isin(folders_main)]
train_df.shape, train_main_df.shape

# Build list of files without auxiliry of train, val, test

val_list_file = Path('..') / 'data' / 'train' / 'validation_list.txt'
test_list_file = Path('..') / 'data' / 'train' / 'testing_list.txt'

with open(val_list_file, 'r') as file:
    set_val_all = set(line.strip().replace('/', '\\') for line in file)
with open(test_list_file, 'r') as file:
    set_test_all = set(line.strip().replace('/', '\\') for line in file)

    
root_folder = Path('..') / 'data' / 'train' / 'audio'

lines = []

set_all = set()

for path in root_folder.rglob('*'):
    if '_background_noise_' not in path.parts and path.is_file():
        set_all.add(str(path.relative_to(root_folder)))
        
set_train_all = set_all - set_val_all - set_test_all
        
len(set_all), len(set_val_all), len(set_test_all), len(set_train_all)

folders_main = ['on', 'four', 'yes', 'nine', 'three', 'no', 'stop', 'five', 'right', 'down', 'one', 'eight', 'off', 'zero', 'two', 'up', 'six', 'go', 'left', 'seven']
print(folders_main)

set_not_main = set()
folders_main_ser = pd.Series(list(folders_main))

for path in root_folder.rglob('*'):
    if not folders_main_ser.isin(path.parts).any() and path.is_file():
        set_not_main.add(str(path.relative_to(root_folder)))
        
set_val_main = set_val_all - set_not_main
set_test_main = set_test_all - set_not_main
set_train_main = set_train_all - set_not_main
        
len(set_train_main), len(set_val_main), len(set_test_main), len(set_not_main), len(set_all)
