==> notebooks/6_mfcc_baseline.ipynb <==
# ! pip install --user librosa

from pathlib import Path
from scipy.io import wavfile
import scipy.signal
import pandas as pd
from tqdm.auto import tqdm
import seaborn as sns
import matplotlib.pyplot as plt
from collections import Counter
import numpy as np
import os
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, classification_report
from sklearn.ensemble import RandomForestClassifier
import librosa

## Read data

intermediate_folder = Path('..') / 'data' / 'intermediate'

X_train = np.load(intermediate_folder / 'train_main_1_sec_audio_mfcc.npy')
X_train.shape

X_val = np.load(intermediate_folder / 'val_main_1_sec_audio_mfcc.npy')
X_val.shape

X_test = np.load(intermediate_folder / 'test_main_1_sec_audio_mfcc.npy')
X_test.shape

y_train = pd.read_csv(intermediate_folder / 'train_main_1_sec_labels.csv', header=None, index_col=False)[0]
y_val = pd.read_csv(intermediate_folder / 'val_main_1_sec_labels.csv', header=None, index_col=False)[0]
y_test = pd.read_csv(intermediate_folder / 'test_main_1_sec_labels.csv', header=None, index_col=False)[0]
y_train.shape, y_val.shape, y_test.shape

X_train.shape, X_train.reshape(X_train.shape[0], -1).shape

## Flat LogisticRegression

lr_def = LogisticRegression(verbose=1, max_iter=300)
lr_def.fit(X_train.reshape(X_train.shape[0], -1), y_train)
y_pred_lr = lr_def.predict(X_val.reshape(X_val.shape[0], -1))
print(classification_report(y_val, y_pred_lr))

lr_def = LogisticRegression(verbose=1, max_iter=1000)
lr_def.fit(X_train.reshape(X_train.shape[0], -1), y_train)
y_pred_lr = lr_def.predict(X_val.reshape(X_val.shape[0], -1))
print(classification_report(y_val, y_pred_lr))

## Flat Random Forest

rf = RandomForestClassifier(n_jobs=-1, verbose=1, min_impurity_decrease=.001)
rf.fit(X_train.reshape(X_train.shape[0], -1), y_train)
y_pred_rf = rf.predict(X_val.reshape(X_val.shape[0], -1))
print(classification_report(y_val, y_pred_rf))

rf = RandomForestClassifier(n_jobs=-1, verbose=1, min_impurity_decrease=.00015)
rf.fit(X_train.reshape(X_train.shape[0], -1), y_train)
y_pred_rf = rf.predict(X_val.reshape(X_val.shape[0], -1))
print(classification_report(y_val, y_pred_rf))

rf = RandomForestClassifier(n_jobs=-1, verbose=1, min_impurity_decrease=.0001)
rf.fit(X_train.reshape(X_train.shape[0], -1), y_train)
y_pred_rf = rf.predict(X_val.reshape(X_val.shape[0], -1))
print(classification_report(y_val, y_pred_rf))

rf = RandomForestClassifier(n_jobs=-1, verbose=1, min_impurity_decrease=.00005)
rf.fit(X_train.reshape(X_train.shape[0], -1), y_train)
y_pred_rf = rf.predict(X_val.reshape(X_val.shape[0], -1))
print(classification_report(y_val, y_pred_rf))

rf = RandomForestClassifier(n_jobs=-1, verbose=1, min_impurity_decrease=.00001)
rf.fit(X_train.reshape(X_train.shape[0], -1), y_train)
y_pred_rf = rf.predict(X_val.reshape(X_val.shape[0], -1))
print(classification_report(y_val, y_pred_rf))

rf = RandomForestClassifier(n_jobs=-1, verbose=1, min_impurity_decrease=.000001)
rf.fit(X_train.reshape(X_train.shape[0], -1), y_train)
y_pred_rf = rf.predict(X_val.reshape(X_val.shape[0], -1))
print(classification_report(y_val, y_pred_rf))
