==> notebooks/9_read_raw_output_stft.ipynb <==
# ! pip install --user librosa

from pathlib import Path
from scipy.io import wavfile
import scipy.signal
import pandas as pd
from tqdm.auto import tqdm
import seaborn as sns
import matplotlib.pyplot as plt
from collections import Counter
import numpy as np
import os
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, classification_report
from sklearn.ensemble import RandomForestClassifier
import librosa

## Read data

intermediate_folder = Path('..') / 'data' / 'intermediate'

X_train = np.load(intermediate_folder / 'train_main_1_sec_audio.npy').astype(float)
X_train.shape

X_train

X_train.mean(axis=1).shape

X_train.mean(axis=1)

X_val = np.load(intermediate_folder / 'val_main_1_sec_audio.npy').astype(float)
X_val.shape

X_test = np.load(intermediate_folder / 'test_main_1_sec_audio.npy').astype(float)
X_test.shape

SAMPLE_RATE = 16000

## Convert and Save librosa

# Compute STFT
stft = librosa.stft(X_train[0])
print(f'Sample shape: {X_train[0].shape}, initial stft {stft.shape=}, {stft.size=}')

# To ensure all the values are real, get magnitude
stft_magnitude, stft_phase = librosa.magphase(stft)
print(f'Before transpose: {stft_magnitude.shape=}')

# Transpose the result so the rows represent time
stft_magnitude = stft_magnitude.T
print(f'After transpose: {stft_magnitude.shape=}')

# To use this data with a Conv2D layer, it needs to have 4 dimensions: (num_samples, rows, columns, channels)
# If we're only using the magnitude (a common choice), we add an extra dimension for channels
stft_data = np.expand_dims(stft_magnitude, axis=-1)
print(f'After adding additional dimension {stft_data.shape=}')

def get_stft_librosa(raw_audio):
    # Compute STFT
    stft_magnitude, stft_phase = librosa.magphase(librosa.stft(raw_audio))
    return stft_magnitude

stft_train_list = []

for i in tqdm(range(len(X_train))):
    stft_train_list.append(get_stft_librosa(X_train[i]))

X_train_stft_np = np.stack(stft_train_list)
del stft_train_list
X_train_stft_np.shape

stft_val_list = []

for i in tqdm(range(len(X_val))):
    stft_val_list.append(get_stft_librosa(X_val[i]))

X_val_stft_np = np.stack(stft_val_list)
del stft_val_list
X_val_stft_np.shape

stft_test_list = []

for i in tqdm(range(len(X_test))):
    stft_test_list.append(get_stft_librosa(X_test[i]))

X_test_stft_np = np.stack(stft_test_list)
del stft_test_list
X_test_stft_np.shape

np.save(intermediate_folder / 'train_main_1_sec_audio_stft_librosa.npy', X_train_stft_np)

np.save(intermediate_folder / 'val_main_1_sec_audio_stft_librosa.npy', X_val_stft_np)

np.save(intermediate_folder / 'test_main_1_sec_audio_stft_librosa.npy', X_test_stft_np)

## Convert and Save Scipy

# Compute the STFT
frequencies, times, Zxx = scipy.signal.stft(X_train[0], fs=SAMPLE_RATE) # , nperseg=128, noverlap=56)

# Note: nperseg (number of samples per segment) and noverlap (number of samples to overlap) 
# can be adjusted according to your needs

# Compute the magnitude (absolute value) of the STFT complex numbers
Zxx = np.abs(Zxx)
frequencies.shape, times.shape, Zxx.shape


# # Plot the spectrogram
# plt.pcolormesh(times, frequencies, 10 * np.log10(Zxx), shading='gouraud')
# plt.title('Spectrogram')
# plt.ylabel('Frequency [Hz]')
# plt.xlabel('Time [sec]')
# plt.show()

def get_stft_scipy(raw_audio):
    frequencies, times, Zxx = scipy.signal.stft(raw_audio, fs=SAMPLE_RATE) # , nperseg=128, noverlap=56)
    return np.abs(Zxx)

stft_train_list = []

for i in tqdm(range(len(X_train))):
    stft_train_list.append(get_stft_scipy(X_train[i]))

X_train_stft_np = np.stack(stft_train_list)
del stft_train_list
X_train_stft_np.shape

stft_val_list = []

for i in tqdm(range(len(X_val))):
    stft_val_list.append(get_stft_scipy(X_val[i]))

X_val_stft_np = np.stack(stft_val_list)
del stft_val_list
X_val_stft_np.shape

stft_test_list = []

for i in tqdm(range(len(X_test))):
    stft_test_list.append(get_stft_scipy(X_test[i]))

X_test_stft_np = np.stack(stft_test_list)
del stft_test_list
X_test_stft_np.shape

np.save(intermediate_folder / 'train_main_1_sec_audio_stft_scipy.npy', X_train_stft_np)

np.save(intermediate_folder / 'val_main_1_sec_audio_stft_scipy.npy', X_val_stft_np)

np.save(intermediate_folder / 'test_main_1_sec_audio_stft_scipy.npy', X_test_stft_np)
