==> notebooks/11_sftf_librosa_CNN.ipynb <==
# ! pip install --user librosa

from pathlib import Path
from scipy.io import wavfile
import scipy.signal
import pandas as pd
from tqdm.auto import tqdm
import seaborn as sns
import matplotlib.pyplot as plt
from collections import Counter
import numpy as np
import os
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, classification_report
from sklearn.ensemble import RandomForestClassifier
import tensorflow as tf
from tensorflow.keras.layers import LSTM, Dense, Bidirectional, Dropout, Conv2D, MaxPooling2D, Flatten, BatchNormalization, Activation
from tensorflow.keras.models import Sequential
from sklearn.preprocessing import LabelEncoder
from tensorflow.keras.optimizers import Adam
from tensorflow.keras.callbacks import EarlyStopping
from tensorflow.keras.regularizers import l1, l2

## Mount Drive

def is_running_on_colab():
    try:
        import google.colab
        return True
    except ImportError:
        return False

ON_COLAB = is_running_on_colab()
ON_COLAB

if ON_COLAB:
  from google.colab import drive
  drive.mount('/content/gdrive')
  intermediate_folder = Path('/content/gdrive/MyDrive/Colab Notebooks/Speech recognition')
  # intermediate_folder = Path('/content/gdrive/MyDrive/Temp/Speech recognition project')
else:
  intermediate_folder = Path('..') / 'data' / 'intermediate'

## Read data

X_train = np.load(intermediate_folder / 'train_main_1_sec_audio_stft_librosa.npy').transpose(0, 2, 1)
X_train.shape

X_val = np.load(intermediate_folder / 'val_main_1_sec_audio_stft_librosa.npy').transpose(0, 2, 1)
X_val.shape

# X_test = np.load(intermediate_folder / 'test_main_1_sec_audio_stft_librosa.npy').transpose(0, 2, 1)
# X_test.shape

y_train_labels = pd.read_csv(intermediate_folder / 'train_main_1_sec_labels.csv', header=None, index_col=False)[0]
y_val_labels = pd.read_csv(intermediate_folder / 'val_main_1_sec_labels.csv', header=None, index_col=False)[0]
# y_test_labels = pd.read_csv(intermediate_folder / 'test_main_1_sec_labels.csv', header=None, index_col=False)[0]
y_train_labels.shape, y_val_labels.shape
# y_train_labels.shape, y_val_labels.shape, y_test_labels.shape

le = LabelEncoder()
le.fit(y_train_labels)
y_train = le.transform(y_train_labels)
y_val = le.transform(y_val_labels)
# y_test = le.transform(y_test_labels)
y_train.shape, y_val.shape
# y_train.shape, y_val.shape, y_test.shape

y_train_labels.value_counts().sort_index()

pd.Series(y_train).value_counts().sort_index()

X_val.shape

assert np.mean(X_val[:, :, 0]) == np.mean(X_val.transpose(0, 2, 1)[:, 0, :])

y_train[:10]

indxs_X_train = np.random.permutation(len(X_train))
X_train = X_train[indxs_X_train]
y_train = y_train[indxs_X_train]
y_train[:10]

## Standardize

MEAN = X_train.mean()
STD = np.std(X_train)
MEAN, STD

X_train = (X_train - MEAN) / STD
X_val = (X_val - MEAN) / STD
# X_test = (X_test - MEAN) / STD
X_train.shape, X_val.shape
# X_train.shape, X_val.shape, X_test.shape

## Functions

def train(model, patience_train, patience_val=None, learning_rate=.001, X_train=X_train, y_train=y_train, X_val=X_val, y_val=y_val, epochs=20, batch_size=128, cont_train=False):
  print(f'{patience_train=}, {patience_val=}, {learning_rate=}, {epochs=}, {batch_size=}, {cont_train=}, {X_train.shape=}, {y_train.shape=}, {X_val.shape=}, {y_val.shape=}')
  model.summary()
  print(f'Input shape {model.input_shape}, output shape {model.output_shape}')
  early_stopping = []
  if patience_train:
    early_stopping.append(EarlyStopping(monitor='loss', patience=patience_train))
  if patience_val:
    early_stopping.append(EarlyStopping(monitor='loss', patience=patience_val))

  if not cont_train:
    model.compile(optimizer=Adam(learning_rate=learning_rate),
                  loss='sparse_categorical_crossentropy',
                  metrics=['accuracy'])
  model.fit(X_train, y_train, batch_size=batch_size, epochs=epochs, validation_data=(X_val, y_val), callbacks=[early_stopping]);

## CNN

model = Sequential([Conv2D(32, (2, 2), activation='relu', padding='same', input_shape=(32, 1025, 1)),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Flatten(),
                    # Dropout(.2),
                    # Dense(32, activation='sigmoid'),
                    # Dropout(.5),
                    # Dense(128, activation='sigmoid'),
                    # Dropout(.2),
                    Dense(20, activation='softmax')])

train(model=model,
      patience_train=10,
      patience_val=None,
      learning_rate=.001,
      epochs=1000,
      batch_size=256,
      # cont_train=True,
      )

model = Sequential([Conv2D(32, (2, 2), activation='relu', padding='same', input_shape=(32, 1025, 1)),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Flatten(),
                    # Dropout(.2),
                    # Dense(32, activation='sigmoid'),
                    # Dropout(.5),
                    # Dense(128, activation='sigmoid'),
                    # Dropout(.2),
                    Dense(20, activation='softmax')])

train(model=model,
      patience_train=10,
      patience_val=None,
      learning_rate=.001,
      epochs=1000,
      batch_size=256,
      # cont_train=True,
      )

model = Sequential([Conv2D(32, (2, 2), activation='relu', padding='same', input_shape=(32, 1025, 1)),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Flatten(),
                    # Dropout(.2),
                    # Dense(32, activation='sigmoid'),
                    # Dropout(.5),
                    # Dense(128, activation='sigmoid'),
                    # Dropout(.2),
                    Dense(20, activation='softmax')])

train(model=model,
      patience_train=3,
      patience_val=None,
      learning_rate=.01,
      epochs=1000,
      batch_size=512,
      # cont_train=True,
      )

model = Sequential([Conv2D(32, (2, 4), activation='relu', padding='same', input_shape=(32, 1025, 1)),
                    MaxPooling2D((2, 4)),
                    Conv2D(32, (2, 4), activation='relu', padding='same'),
                    MaxPooling2D((2, 4)),
                    Conv2D(32, (2, 4), activation='relu', padding='same'),
                    MaxPooling2D((2, 4)),
                    Conv2D(32, (2, 4), activation='relu', padding='same'),
                    MaxPooling2D((2, 4)),
                    Flatten(),
                    # Dropout(.2),
                    # Dense(32, activation='sigmoid'),
                    # Dropout(.5),
                    # Dense(128, activation='sigmoid'),
                    # Dropout(.2),
                    Dense(20, activation='softmax')])

train(model=model,
      patience_train=3,
      patience_val=None,
      learning_rate=.01,
      epochs=1000,
      batch_size=512,
      # cont_train=True,
      )

model = Sequential([Conv2D(32, (3, 6), activation='relu', padding='same', input_shape=(32, 1025, 1)),
                    MaxPooling2D((3, 6)),
                    Conv2D(32, (3, 6), activation='relu', padding='same'),
                    MaxPooling2D((3, 6)),
                    Conv2D(32, (3, 6), activation='relu', padding='same'),
                    MaxPooling2D((3, 6)),
                    Flatten(),
                    # Dropout(.2),
                    # Dense(32, activation='sigmoid'),
                    # Dropout(.5),
                    # Dense(128, activation='sigmoid'),
                    # Dropout(.2),
                    Dense(20, activation='softmax')])

train(model=model,
      patience_train=3,
      patience_val=None,
      learning_rate=.01,
      epochs=1000,
      batch_size=512,
      # cont_train=True,
      )

model = Sequential([Conv2D(32, (3, 6), activation='relu', padding='same', input_shape=(32, 1025, 1)),
                    MaxPooling2D((3, 6)),
                    Conv2D(64, (3, 6), activation='relu', padding='same'),
                    MaxPooling2D((3, 6)),
                    Conv2D(128, (3, 6), activation='relu', padding='same'),
                    MaxPooling2D((3, 6)),
                    Flatten(),
                    # Dropout(.2),
                    # Dense(128, activation='relu'),
                    # Dropout(.5),
                    Dense(20, activation='softmax')])

train(model=model,
      patience_train=3,
      patience_val=None,
      learning_rate=.01,
      epochs=1000,
      batch_size=512,
      # cont_train=True,
      )

model = Sequential([Conv2D(32, (3, 6), activation='relu', padding='same', input_shape=(32, 1025, 1)),
                    MaxPooling2D((3, 6)),
                    Conv2D(64, (3, 6), activation='relu', padding='same'),
                    MaxPooling2D((3, 6)),
                    Conv2D(128, (3, 6), activation='relu', padding='same'),
                    MaxPooling2D((3, 6)),
                    Flatten(),
                    # Dropout(.2),
                    Dense(128, activation='relu'),
                    # Dropout(.5),
                    Dense(20, activation='softmax')])

train(model=model,
      patience_train=3,
      patience_val=None,
      learning_rate=.01,
      epochs=1000,
      batch_size=512,
      # cont_train=True,
      )

model = Sequential([Conv2D(32, (2, 4), activation='relu', padding='same', input_shape=(32, 1025, 1)),
                    MaxPooling2D((2, 4)),
                    Conv2D(64, (2, 4), activation='relu', padding='same'),
                    MaxPooling2D((2, 4)),
                    Conv2D(128, (2, 4), activation='relu', padding='same'),
                    MaxPooling2D((2, 4)),
                    Conv2D(256, (2, 4), activation='relu', padding='same'),
                    MaxPooling2D((2, 4)),
                    Flatten(),
                    # Dropout(.2),
                    # Dense(32, activation='sigmoid'),
                    # Dropout(.5),
                    # Dense(128, activation='sigmoid'),
                    # Dropout(.2),
                    Dense(20, activation='softmax')])

train(model=model,
      patience_train=3,
      patience_val=None,
      learning_rate=.01,
      epochs=1000,
      batch_size=512,
      # cont_train=True,
      )

model = Sequential([Conv2D(32, (2, 4), activation='relu', padding='same', input_shape=(32, 1025, 1)),
                    MaxPooling2D((2, 4)),
                    Conv2D(64, (2, 4), activation='relu', padding='same'),
                    MaxPooling2D((2, 4)),
                    Conv2D(64, (2, 4), activation='relu', padding='same'),
                    MaxPooling2D((2, 4)),
                    Conv2D(64, (2, 4), activation='relu', padding='same'),
                    MaxPooling2D((2, 4)),
                    Flatten(),
                    # Dropout(.2),
                    # Dense(32, activation='sigmoid'),
                    # Dropout(.5),
                    # Dense(128, activation='sigmoid'),
                    # Dropout(.2),
                    Dense(20, activation='softmax')])

train(model=model,
      patience_train=3,
      patience_val=None,
      learning_rate=.01,
      epochs=1000,
      batch_size=512,
      # cont_train=True,
      )

model = Sequential([Conv2D(32, (2, 4), activation='relu', padding='same', input_shape=(32, 1025, 1)),
                    MaxPooling2D((2, 4)),
                    Conv2D(64, (2, 4), activation='relu', padding='same'),
                    MaxPooling2D((2, 4)),
                    Conv2D(128, (2, 4), activation='relu', padding='same'),
                    MaxPooling2D((2, 4)),
                    Conv2D(128, (2, 4), activation='relu', padding='same'),
                    MaxPooling2D((2, 4)),
                    Flatten(),
                    # Dropout(.2),
                    # Dense(32, activation='sigmoid'),
                    # Dropout(.5),
                    # Dense(128, activation='sigmoid'),
                    # Dropout(.2),
                    Dense(20, activation='softmax')])

train(model=model,
      patience_train=3,
      patience_val=None,
      learning_rate=.01,
      epochs=1000,
      batch_size=512,
      # cont_train=True,
      )

model = Sequential([Conv2D(32, (2, 2), activation='relu', padding='same', input_shape=(32, 1025, 1)),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Flatten(),
                    Dropout(.3),
                    # Dense(32, activation='rele'),
                    # Dropout(.5),
                    Dense(20, activation='softmax')])

train(model=model,
      patience_train=3,
      patience_val=None,
      learning_rate=.01,
      epochs=1000,
      batch_size=512,
      # cont_train=True,
      )

## End
