==> notebooks/20_mfcc_cnn_test.ipynb <==
import importlib

if importlib.util.find_spec('librosa') is None:
  ! pip install --user librosa
else:
  print('libraries are already installed')

from pathlib import Path
# from scipy.io import wavfile
# import scipy.signal
import pandas as pd
from tqdm.auto import tqdm
# import seaborn as sns
# import matplotlib.pyplot as plt
# from collections import Counter
import numpy as np
# import os
# from sklearn.linear_model import LogisticRegression
# from sklearn.metrics import accuracy_score, classification_report
# from sklearn.ensemble import RandomForestClassifier
import tensorflow as tf
from tensorflow.keras.layers import LSTM, Dense, Bidirectional, Dropout, Conv2D, MaxPooling2D, Flatten, BatchNormalization, Reshape
from tensorflow.keras.models import Sequential
# from sklearn.preprocessing import LabelEncoder
# from tensorflow.keras.optimizers import Adam
# from tensorflow.keras.callbacks import EarlyStopping
# from tensorflow.keras.regularizers import l1, l2
# from transformers import AutoProcessor, TFWav2Vec2Model
# from tensorflow.keras import mixed_precision
from sklearn.metrics import classification_report
import json
import librosa

## Mount Drive

def is_running_on_colab():
    try:
        import google.colab
        return True
    except ImportError:
        return False

ON_COLAB = is_running_on_colab()
ON_COLAB

if ON_COLAB:
  from google.colab import drive
  drive.mount('/content/gdrive')
  intermediate_folder = Path('/content/gdrive/MyDrive/Colab Notebooks/Speech recognition')
  # intermediate_folder = Path('/content/gdrive/MyDrive/Temp/Speech recognition project')
else:
  intermediate_folder = Path('..') / 'data' / 'intermediate'

## Read and define model

model = Sequential([Conv2D(32, (2, 2), activation='relu', padding='same', input_shape=(32, 20, 1)),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Conv2D(32, (2, 2), activation='relu', padding='same'),
                    MaxPooling2D((2, 2)),
                    Flatten(),
                    # Dropout(.2),
                    # Dense(32, activation='relu'),
                    Dropout(.175),
                    Dense(20, activation='softmax')])
model.summary()

model.load_weights(intermediate_folder / 'best_model_mfcc_cnn.ckp')

## Read data

SAMPLE_RATE = 16000

BATCH_SIZE = 256

# X_train = np.load(intermediate_folder / 'train_main_1_sec_audio.npy').astype(float)
# X_train.shape

# mfcc_train_list = []

# for i in tqdm(range(len(X_train))):
#     mfcc_train_list.append(librosa.feature.mfcc(y=X_train[i], sr=SAMPLE_RATE))

# X_train_mfcc_np = np.stack(mfcc_train_list)
# del mfcc_train_list
# X_train_mfcc_np.shape

X_val = np.load(intermediate_folder / 'val_main_1_sec_audio.npy').astype(float)
X_val.shape

mfcc_val_list = []

for i in tqdm(range(len(X_val))):
    mfcc_val_list.append(librosa.feature.mfcc(y=X_val[i], sr=SAMPLE_RATE))

X_val_mfcc_np = np.stack(mfcc_val_list)
del mfcc_val_list
X_val_mfcc_np.shape

X_test = np.load(intermediate_folder / 'test_main_1_sec_audio.npy').astype(float)
X_test.shape

mfcc_test_list = []

for i in tqdm(range(len(X_test))):
    mfcc_test_list.append(librosa.feature.mfcc(y=X_test[i], sr=SAMPLE_RATE))

X_test_mfcc_np = np.stack(mfcc_test_list)
del mfcc_test_list
X_test_mfcc_np.shape

# y_train_labels = pd.read_csv(intermediate_folder / 'train_main_1_sec_labels.csv', header=None, index_col=False)[0]
# y_train_labels.shape

y_val_labels = pd.read_csv(intermediate_folder / 'val_main_1_sec_labels.csv', header=None, index_col=False)[0]
y_val_labels.shape

y_test_labels = pd.read_csv(intermediate_folder / 'test_main_1_sec_labels.csv', header=None, index_col=False)[0]
y_test_labels.shape

with open(intermediate_folder / 'labels.json', 'r') as file:
    labels_dict = json.load(file)
labels_dict

# le = LabelEncoder()
# le.fit(y_train_labels)

# y_train = le.transform(y_train_labels)
# y_val = le.transform(y_val_labels)
# y_test = le.transform(y_test_labels)

# del y_test_labels
# del y_train_labels
# del y_val_labels
# y_test.shape

# y_train = np.array([labels_dict[lbl] for lbl in y_train_labels])
y_val = np.array([labels_dict[lbl] for lbl in y_val_labels])
y_test = np.array([labels_dict[lbl] for lbl in y_test_labels])

# del y_train_labels
del y_val_labels
del y_test_labels

# import json
# classes = {cl: i for i, cl in enumerate(le.classes_)}
# print(classes)
# with open(intermediate_folder / 'labels.json', 'w') as fp:
#     json.dump(classes, fp)

## Standardize

MEAN = 33.06522526955586
STD = 142.73227249991973
MEAN, STD

# X_train_mfcc_np = (X_train - MEAN) / STD
X_val_mfcc_np = (X_val_mfcc_np - MEAN) / STD
X_test_mfcc_np = (X_test_mfcc_np - MEAN) / STD
X_val_mfcc_np.shape, X_test_mfcc_np.shape
# X_train.shape, X_val.shape, X_test.shape

## Predict and metrics

# # y_pred = model.predict(np.expand_dims(X_train_mfcc_np, axis=-1), batch_size=BATCH_SIZE)
# y_pred = model.predict(np.expand_dims(X_train_mfcc_np.transpose(0, 2, 1), axis=-1), batch_size=BATCH_SIZE)
# print((y_pred.argmax(axis=1) == y_train).mean())
# print(classification_report(y_train, y_pred.argmax(axis=1)))

# y_pred = model.predict(np.expand_dims(X_val_mfcc_np, axis=-1), batch_size=BATCH_SIZE)
y_pred = model.predict(np.expand_dims(X_val_mfcc_np.transpose(0, 2, 1), axis=-1), batch_size=BATCH_SIZE)
print((y_pred.argmax(axis=1) == y_val).mean())
print(classification_report(y_val, y_pred.argmax(axis=1)))

# y_pred = model.predict(np.expand_dims(X_test_mfcc_np, axis=-1), batch_size=BATCH_SIZE)
y_pred = model.predict(np.expand_dims(X_test_mfcc_np.transpose(0, 2, 1), axis=-1), batch_size=BATCH_SIZE)
print((y_pred.argmax(axis=1) == y_test).mean())
print(classification_report(y_test, y_pred.argmax(axis=1)))
