==> notebooks/31_gradio_with_wav2vec_and_mfcc_cnn.ipynb <==
import importlib

if importlib.util.find_spec('librosa') is None:
  ! pip install --user librosa
else:
  print('librosa already installed')

if importlib.util.find_spec('gradio') is None:
  ! pip install gradio
else:
  print('gradio already installed')

if importlib.util.find_spec('transformers') is None:
    print('Installing transformers...')
    ! pip install transformers
    ! pip install pyctcdecode
    ! pip install https://github.com/kpu/kenlm/archive/master.zip
else:
    print('libraries are already installed')

from pathlib import Path
# from scipy.io import wavfile
# import scipy.signal
import pandas as pd
from tqdm.auto import tqdm
# import seaborn as sns
# import matplotlib.pyplot as plt
# from collections import Counter
import numpy as np
# import os
# from sklearn.linear_model import LogisticRegression
# from sklearn.metrics import accuracy_score, classification_report
# from sklearn.ensemble import RandomForestClassifier
import tensorflow as tf
from tensorflow.keras.layers import LSTM, Dense, Bidirectional, Dropout, Conv2D, MaxPooling2D, Flatten, BatchNormalization, Reshape
from tensorflow.keras.models import Sequential
# from sklearn.preprocessing import LabelEncoder
# from tensorflow.keras.optimizers import Adam
# from tensorflow.keras.callbacks import EarlyStopping
# from tensorflow.keras.regularizers import l1, l2
from transformers import AutoProcessor, TFWav2Vec2Model
# from tensorflow.keras import mixed_precision
from sklearn.metrics import classification_report
import json
import librosa
import gradio as gr

## Mount Drive

def is_running_on_colab():
    try:
        import google.colab
        return True
    except ImportError:
        return False

ON_COLAB = is_running_on_colab()
ON_COLAB

if ON_COLAB:
  from google.colab import drive
  drive.mount('/content/gdrive')
  intermediate_folder = Path('/content/gdrive/MyDrive/Colab Notebooks/Speech recognition')
  # intermediate_folder = Path('/content/gdrive/MyDrive/Temp/Speech recognition project')
else:
  intermediate_folder = Path('..') / 'data' / 'intermediate'

## Read and define model

SAMPLE_RATE = 16000

model_cnn = Sequential([Conv2D(32, (2, 2), activation='relu', padding='same', input_shape=(32, 20, 1)),
                        MaxPooling2D((2, 2)),
                        Conv2D(32, (2, 2), activation='relu', padding='same'),
                        MaxPooling2D((2, 2)),
                        Conv2D(32, (2, 2), activation='relu', padding='same'),
                        MaxPooling2D((2, 2)),
                        Flatten(),
                        # Dropout(.2),
                        # Dense(32, activation='relu'),
                        Dropout(.175),
                        Dense(20, activation='softmax')])
model_cnn.load_weights(intermediate_folder / 'best_model_mfcc_cnn.ckp')
model_cnn.summary()

model_name = 'facebook/wav2vec2-base-960h'
processor = AutoProcessor.from_pretrained(model_name)
model_wav2vec_hf = TFWav2Vec2Model.from_pretrained(model_name)

model_wav2vec_finetune = Sequential([
    Conv2D(4, (49, 1), activation='relu', padding='same', input_shape=(49, 768, 1)),
    MaxPooling2D((49, 1)),
    # Conv2D(2, (3, 1), activation='relu', padding='same'),
    # MaxPooling2D((3, 1)),
    # Conv2D(2, (3, 1), activation='relu', padding='same'),
    # MaxPooling2D((3, 1)),
    Flatten(),
    Dropout(.5),
    # Dense(32, activation='relu'),
    # Dropout(.5),
    # Dense(128, activation='relu'),
    # Dropout(.2),
    Dense(20, activation='softmax')])

input_ids = tf.keras.Input(shape=(SAMPLE_RATE,), dtype='float64', name='input')
attention_masks = tf.keras.Input(shape=(SAMPLE_RATE,), dtype='int32', name='mask')

output = model_wav2vec_hf([input_ids, attention_masks])
output = output[0]
output = model_wav2vec_finetune(output)

model_wav2vec_full = tf.keras.models.Model(inputs={"input": input_ids, "mask": attention_masks}, outputs=output)
model_wav2vec_full.load_weights(intermediate_folder / 'full_model_weights.h5')
model_wav2vec_full.summary()

## Read data

X_test = np.load(intermediate_folder / 'test_main_1_sec_audio.npy')
# X_test = np.load(intermediate_folder / 'test_main_1_sec_audio.npy').astype(float)
X_test.shape

y_test_labels = pd.read_csv(intermediate_folder / 'test_main_1_sec_labels.csv', header=None, index_col=False)[0]
y_test_labels.shape

with open(intermediate_folder / 'labels.json', 'r') as file:
    labels_dict = json.load(file)
labels_dict

# y_train = np.array([labels_dict[lbl] for lbl in y_train_labels])
# y_val = np.array([labels_dict[lbl] for lbl in y_val_labels])
y_test = np.array([labels_dict[lbl] for lbl in y_test_labels])

# del y_train_labels
# del y_val_labels
del y_test_labels

## Function of predict

MEAN = 33.06522526955586
STD = 142.73227249991973
MEAN, STD

SAMPLE_RATE = 16000

labels_dict_reversed = {v:k for k, v in labels_dict.items()}
labels_dict_reversed

def predict_with_expected(data, expected):
  data_transformed = np.expand_dims(np.expand_dims(librosa.feature.mfcc(y=data.astype(float), sr=SAMPLE_RATE).transpose(1, 0), axis=-1), axis=0)
  data_transformed = (data_transformed - MEAN) / STD
  y_pred = model_cnn.predict(data_transformed, verbose=0)
  print(f'Largest class: {y_pred.argmax(axis=1)[0]}, expected: {expected}, same: {y_pred.argmax(axis=1)[0] == expected}')

for ind in np.random.randint(len(X_test), size=10):
  predict_with_expected(X_test[ind], y_test[ind])

def softmax(x):
    """Compute softmax values for each sets of scores in x."""
    e_x = np.exp(x - np.max(x))
    return e_x / e_x.sum(axis=0)

num_used = 0

def predict(option, data_mic, data_upload):
  global num_used
  num_used += 1
  print(f'Being used {num_used} time!')
  data = data_mic if data_mic else data_upload
  if data:
    sample_rate = data[0]
    audio = data[1]
    orig_audio = audio
  else:
    return "", None, None

#   print(f'---------------------')
  if sample_rate != SAMPLE_RATE:
    # print(f'Before resampling {sample_rate=}, {len(audio)=}')
    audio = librosa.resample(y=audio.astype(float), orig_sr=sample_rate, target_sr=SAMPLE_RATE)
  else:
    # print(f'Sample rate as expected {sample_rate=}, {len(audio)=}')
    audio = audio.astype(np.float32)

  # print(f'Before {SAMPLE_RATE=}, {len(audio)=}')
  if len(audio) > SAMPLE_RATE:
    audio, inds = librosa.effects.trim(audio)
    # print(f'After trimming {SAMPLE_RATE=}, {len(audio)=}, {inds=}')

  # if len(audio) > SAMPLE_RATE:
  #   audio = audio[:SAMPLE_RATE]
  #   print(f'After removing end {SAMPLE_RATE=}, {len(audio)=}')

  if len(audio) > SAMPLE_RATE:
    # print(f'Finding the loudest second')
    frames = librosa.util.frame(audio,
                                frame_length=SAMPLE_RATE,
                                hop_length=SAMPLE_RATE // 10,
                                axis=0
                                )
    # print(f'{len(frames)=}, {frames.shape=}')

    # Calculate the RMS (root mean square) of each frame
    rms = np.sqrt(np.mean(np.abs(frames)**2, axis=1))
    # print(f'{rms.shape=} {rms=}')

    # Find the index of the frame with the highest RMS
    loudest_frame_index = np.argmax(rms)
    # print(f'Found loudest frame {loudest_frame_index} from total {len(rms)}')

    # Get the audio for this frame
    audio = frames[loudest_frame_index, :]

  if len(audio) < SAMPLE_RATE:
    # print(f'Padding')
    audio = np.pad(audio, (0, SAMPLE_RATE - len(audio)))

  # print(f'After putting as 1 second {SAMPLE_RATE=}, {len(audio)=}')

  if option == 'CNN-based' or option is None:
    data_transformed = np.expand_dims(np.expand_dims(librosa.feature.mfcc(y=audio, sr=SAMPLE_RATE).transpose(1, 0), axis=-1), axis=0)
    data_transformed = (data_transformed - MEAN) / STD
    # print(f'Shape on input to model {data_transformed.shape=}')
    y_pred = model_cnn.predict(data_transformed, verbose=0)
    # print(f'{y_pred.shape=}')
  else:
    data_transformed = processor(audio, return_tensors="tf", sampling_rate=SAMPLE_RATE).input_values
    # print(data_transformed.shape)
    y_pred = model_wav2vec_full.predict([data_transformed, tf.ones([1, SAMPLE_RATE], dtype=tf.int32)], verbose=0)

  bar_data = pd.DataFrame({'Probability': softmax(y_pred[0]).round(2),
                           'Word': [labels_dict_reversed[ind] for ind in range(len(labels_dict_reversed))]}
                           ).sort_values(by='Probability', ascending=False)
  # print(bar_data)
  if audio.shape == orig_audio.shape and (audio == orig_audio).all():
    res_audio = None
  else:
    res_audio = (SAMPLE_RATE, audio)

  return labels_dict_reversed[y_pred.argmax(axis=1)[0]], bar_data.iloc[:5] , res_audio

## Gradio

demo = gr.Interface(fn=predict,
                    inputs=[gr.components.Radio(['CNN-based', 'wav2vec-based'], value='CNN-based', label='Choose model type. If not chosen, CNN-based model will be used.', show_label=True),
                            # gr.components.Dropdown(['CNN-based', 'wav2vec-based'], value='CNN-based', label='Choose model type', show_label=True),
                            gr.components.Audio(source="microphone", autoplay=True, label='Record a few seconds of audio with a short (< 1 sec) clear word in the middle', show_label=True),
                            gr.components.Audio(source="upload", autoplay=True, label='Or upload an audio file with one of the words', show_label=True)],
                    outputs=[gr.components.Textbox(label='Result - most probable word:', show_label=True),
                             gr.components.BarPlot(x='Word', y='Probability', vertical=False, label='Probas', y_lim=[0, 1], width=300, show_label=True),
                             gr.components.Audio(label='Transformed audio predicted on (sampling change, trimmed, padded, loudest second chosen etc.)', show_label=True, autoplay=False)
                             ],
                    live=True,
                    allow_flagging='never',
                    title='Audio word classification',
                    description="""<h3>Record or upload a file with audio with one of the following words:</h3>
                    <table>
                      <tr>
                        <td valign='top'>
                          <h3>Numbers:</h3>
                          <p>zero, one, two, three, four, five, six, seven, eight, nine</p>
                        </td>
                        <td valign='top'>
                          <h3>Additional words:</h3>
                          <p>yes, no, down, up, left, right, go, stop, on, off</p>
                        </td>
                      </tr>
                    </table>
                    2 options for models are given: Simpler CNN-based model (14K params, accuracy 88%), HF wav2vec-based model (94M params, accuracy 95%).
                    <br>You can download audio file examples <a href="https://drive.google.com/drive/folders/1JhyPQ6Zl2FxwDMx1jCwpXhqt5jHzOr-t">from here</a>
                    <br>If the audio is both recorded and uploaded, the recorded audio is used.""")

demo.launch(debug=True, share=True)
